requirements.txt
requests
bs4
docx
xmltodict
streamlit
python-docx
tqdm
pytube
python-dotenv
ffmpeg
azure-cognitiveservices-speech
openpyxl
openai
streamlit_chat
flask
opencv-python
PyMuPDF